// Frontend/src/app/api/models/application-list-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ApplicationListDto {
  candidateEmail?: string;
  candidateName?: string;
  cvUrl?: string;
  highestDegree?: string;
  id: number;
  majorField?: string;
  scoreIA?: number;
  status?: 'DRAFT' | 'SUBMITTED' | 'UNDER_REVIEW' | 'AI_SCORED' | 'SHORTLISTED' | 'INTERVIEW_SCHEDULED' | 'INTERVIEW_COMPLETED' | 'ACCEPTED' | 'REJECTED' | 'WITHDRAWN';
  submitDate?: string;
}